Move `MockServerVerifier` to tests::common
